requirements.txt
# Requires Python 3.11

python-dotenv>=1.0.0
psycopg2-binary>=2.9.10
beautifulsoup4>=4.12.0
requests>=2.31.0
